+++ 2016AD_G04_Servidor/src/main/java/com/ADG04/Servidor/dao/ProductoEncomiendaDao.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 
-public class ProductoEncomiendaDao extends GenericDao<ProductoEncomienda, Long> {
+public class ProductoEncomiendaDao extends GenericDao<ProductoEncomienda, Integer> {
 
 	
 	public ProductoEncomiendaDao(EntityManager entityManager) {
