+++ 2016AD_G04_Servidor/src/main/java/com/ADG04/Servidor/dao/RemitoDao.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 
-public class RemitoDao extends GenericDao<Remito, Long> {
+public class RemitoDao extends GenericDao<Remito, Integer> {
 
 	
 	public RemitoDao(EntityManager entityManager) {
