+++ 2016AD_G04_Servidor/src/main/java/com/ADG04/Servidor/dao/TipoProveedorDao.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 
-public class TipoProveedorDao extends GenericDao<TipoProveedor, Long> {
+public class TipoProveedorDao extends GenericDao<TipoProveedor, Integer> {
 
 	
 	public TipoProveedorDao(EntityManager entityManager) {
