+++ 2016AD_G04_Servidor/src/main/java/com/ADG04/Servidor/dao/VehiculoDao.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 
-public class VehiculoDao extends GenericDao<Vehiculo, Long> {
+public class VehiculoDao extends GenericDao<Vehiculo, Integer> {
 
 	
 	public VehiculoDao(EntityManager entityManager) {
